tools/spt/requirements.txt
# Speech Processing Test (SPT) Dependencies

# Google Gemini API
google-generativeai>=0.8.0

# Environment variables
python-dotenv>=1.0.0

# Audio processing
pydub>=0.25.1

# CLI interface
click>=8.1.0

# Utilities
requests>=2.31.0

# Web API (for web interface)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
